src/auth/service/requirements-build.txt
asyncpg
cryptography
fastapi
psycopg2
pydantic
pyjwt
sqlalchemy
uvicorn
